Don't add f argument to URL if it's already there
When calling the backend Coverages server, the data store adds
the argument "f=netcdf" to specify the file format. However,
it's possible that this parameter is already included in the
coverage endpoint provided by the OGC API. The store now checks
for the parameter and only adds it if it's missing.